Early import of pending upstream commmit
For Trusty we want to be able to reduce the recursion limit for CBOR
parsing, as the default is bigger than the normal stack limit in Trusty.

This change has been added upstream as commit 25174b7bc31f ("feat: add
from_reader_with_recursion_limit") from
https://github.com/enarx/ciborium/pull/84, so it should be available in
the next release of ciborium.

Test: TreeHugger, local Trusty fuzzing build
Change-Id: Ie1ca9affc26927b2d83722b705d1e4c10200939f